src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/primitive/ArrayPrimitiveIntegerTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.primitive;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayPrimitiveIntegerTextParser extends ArrayTextParser {

    public ArrayPrimitiveIntegerTextParser() {
        super(0);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Integer.parseInt(element.getText());
    }

}
